Phase 2 - implemented JSONReader and JsonWriter methods in GoalTrackerApp

+++ src/main/model/Goal.java
@@ -8,13 +8,13 @@ public class Goal implements Writable {
     private static int nextGoalID = 1;
     private int id;
     private String name;
-    private double progress;
+    private int progress;
 
     // REQUIRES: goalName has nonzero length
     // EFFECTS: goal is set to goalName, goal id is integer not assigned to any other goal
     // if initialProgress >= 0 then progress on account is set to
     // initialProgress, otherwise progress is zero.
-    public Goal(String goalName, double initialProgress) {
+    public Goal(String goalName, int initialProgress) {
         id = nextGoalID++;
         name = goalName;
         if (initialProgress >= 0) {
@@ -42,13 +42,13 @@ public class Goal implements Writable {
     // REQUIRES: amount >= 0
     // MODIFIES: this
     // EFFECTS: amount is added to progress and progress is returned
-    public double addTime(double amount) {
+    public double addTime(int amount) {
         progress = progress + amount;
         return progress;
     }
 
     // EFFECTS: returns progress = to amount
-    public double trackProgress(double amount) {
+    public double trackProgress(int amount) {
         progress = amount;
         return progress;
     }
